requirements.txt
gym==0.21.0
mujoco-py==2.1.2.14
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
higher==0.2.1
cherry-rl==0.1.4
PyVirtualDisplay==3.0
scikit-learn==1.2.0
wandb==0.13.7
opencv-python==4.7.0.68
pandas==1.5.2
stable-baselines3==1.5.0
tqdm==4.64.1
hydra-core==1.3.1
einops==0.6.0
tensorboard==2.11.0
gtimer==1.0.0b5